requirements.txt
streamlit
pymupdf
python-docx
python-pptx
sumy
nltk
fpdf